Pull RefreshTopicMetadata up to interface.
I took the opportunity to clean up the code here as well.
The BrokerRoute class was pulled out into its own file in the Model
directory.
Swapped out the hit miss tuple for a class to make the code a bit more
understandable.
Add a new test ensuring refresh function ignores the cache.