Revert change: buffer overflow was symptom not cause
The real cause of the buffer overflow problem was when a messageset was
returned the got cut off by the maxbytes in a fetch request.  This would
cause the last message in a messageset to be mal-formed.

I am reverting this back to a basic read because the message should
define how much should be returned in the stream.  I am adding an
Available check method which the consumer of the stream object can use
to ensure there is enough space to read a certain byte size.

Preventing a read problem throwing here simply caused the problem to
move up to somewhere else in the code.